// karabiner/types.ts
export interface KarabinerRule {
  description?: string;
  manipulators?: Manipulator[];
}

export interface Manipulator {
  description?: string;
  type: "basic";
  from: From;
  to?: To[];
  to_after_key_up?: To[];
  to_if_alone?: To[];
  to_if_held_down?: To[];
  parameters?: Parameters;
  conditions?: Condition[];
}

export interface Parameters {
  "basic.simultaneous_threshold_milliseconds"?: number;
  "basic.to_if_alone_timeout_milliseconds"?: number,
  "basic.to_if_held_down_threshold_milliseconds"?: number
}

export type Condition = FrontMostApplicationCondition | DeviceCondition | KeybaordTypeCondition | InputSourceCondition | VaribaleCondition | EventChangedCondition;
// ...
export interface From {
  key_code?: KeyCode;
  simultaneous?: SimultaneousFrom[];
  simultaneous_options?: SimultaneousOptions;
  modifiers?: Modifiers;
  lazy?: boolean;
}

export interface Modifiers {
  optional?: string[];
  mandatory?: string[];
}

export interface To {
  key_code?: KeyCode;
  modifiers?: string[];
  shell_command?: string;
  set_variable?: {
    name: string;
    value: boolean | number | string;
  };
  mouse_key?: MouseKey;
  lazy?: boolean;
}